requirements.txt
streamlit==1.39.0
tensorflow==2.17.0
numpy==1.26.4
Pillow==10.4.0
pandas==2.1.4
matplotlib==3.8.4
